// src/core/api/sgp/types.ts
export interface SgpGameSummaryLol {
    metadata: SgpGameMetadataLol
    json: SgpGameSummaryJsonLol
}

export interface SgpGameSummaryJsonLol {
    endOfGameResult: string
    gameCreation: number
    gameDuration: number
    gameEndTimestamp: number
    gameId: number
    gameMode: string
    gameName: string
    gameStartTimestamp: number
    gameType: string
    gameVersion: string
    mapId: number
    participants: SgpParticipantLol[]
    platformId: string
    queueId: number
    seasonId: number
    teams: Team[]
    tournamentCode: string
}

export interface Team {
    bans: Ban[]
    objectives: Objectives
    teamId: number
    win: boolean
}

interface Objectives {
    baron: Baron
    champion: Baron
    dragon: Baron
    horde: Baron
    inhibitor: Baron
    riftHerald: Baron
    tower: Baron
}

interface Baron {
    first: boolean
    kills: number
}

interface Ban {
    championId: number
    pickTurn: number
}

export interface SgpParticipantLol {
    allInPings: number
    assistMePings: number
    assists: number
    baronKills: number
    basicPings: number
    bountyLevel: number
    challenges: Challenges
    champExperience: number
    champLevel: number
    championId: number
    championName: string
    championTransform: number
    commandPings: number
    consumablesPurchased: number
    damageDealtToBuildings: number
    damageDealtToObjectives: number
    damageDealtToTurrets: number
    damageSelfMitigated: number
    dangerPings: number
    deaths: number
    detectorWardsPlaced: number
    doubleKills: number
    dragonKills: number
    eligibleForProgression: boolean
    enemyMissingPings: number
    enemyVisionPings: number
    firstBloodAssist: boolean
    firstBloodKill: boolean
    firstTowerAssist: boolean
    firstTowerKill: boolean
    gameEndedInEarlySurrender: boolean
    gameEndedInSurrender: boolean
    getBackPings: number
    goldEarned: number
    goldSpent: number
    holdPings: number
    individualPosition: string
    inhibitorKills: number
    inhibitorTakedowns: number
    inhibitorsLost: number
    item0: number
    item1: number
    item2: number
    item3: number
    item4: number
    item5: number
    item6: number
    itemsPurchased: number
    killingSprees: number
    kills: number
    lane: string
    largestCriticalStrike: number
    largestKillingSpree: number
    largestMultiKill: number
    longestTimeSpentLiving: number
    magicDamageDealt: number
    magicDamageDealtToChampions: number
    magicDamageTaken: number
    missions: SgpMissions
    needVisionPings: number
    neutralMinionsKilled: number
    nexusKills: number
    nexusLost: number
    nexusTakedowns: number
    objectivesStolen: number
    objectivesStolenAssists: number
    onMyWayPings: number
    participantId: number
    pentaKills: number
    perks: Perks
    physicalDamageDealt: number
    physicalDamageDealtToChampions: number
    physicalDamageTaken: number
    placement: number
    playerAugment1: number
    playerAugment2: number
    playerAugment3: number
    playerAugment4: number
    playerAugment5: number
    playerAugment6: number
    playerSubteamId: number
    profileIcon: number
    pushPings: number
    puuid: string
    quadraKills: number
    riotIdGameName: string
    riotIdTagline: string
    role: string
    sightWardsBoughtInGame: number
    spell1Casts: number
    spell1Id: number
    spell2Casts: number
    spell2Id: number
    spell3Casts: number
    spell4Casts: number
    subteamPlacement: number
    summoner1Casts: number
    summoner2Casts: number
    summonerId: number
    summonerLevel: number
    summonerName: string
    teamEarlySurrendered: boolean
    teamId: number
    teamPosition: string
    timeCCingOthers: number
    timePlayed: number
    totalAllyJungleMinionsKilled: number
    totalDamageDealt: number
    totalDamageDealtToChampions: number
    totalDamageShieldedOnTeammates: number
    totalDamageTaken: number
    totalEnemyJungleMinionsKilled: number
    totalHeal: number
    totalHealsOnTeammates: number
    totalMinionsKilled: number
    totalTimeCCDealt: number
    totalTimeSpentDead: number
    totalUnitsHealed: number
    tripleKills: number
    trueDamageDealt: number
    trueDamageDealtToChampions: number
    trueDamageTaken: number
    turretKills: number
    turretTakedowns: number
    turretsLost: number
    unrealKills: number
    visionClearedPings: number
    visionScore: number
    visionWardsBoughtInGame: number
    wardsKilled: number
    wardsPlaced: number
    win: boolean
}
// ...
export interface SgpGameMetadataLol {
    product: string
    tags: string[]
    participants: string[]
    timestamp: string
    data_version: string
    info_type: string
    match_id: string
    private: boolean
}